third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/word-wrap-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'word-wrap' to the 'break-spaces' keyword: break-spaces
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] 'word-wrap' does not support 'break-word break-spaces'
  assert_class_string: Unsupported value must be a CSSStyleValue and not one of its subclasses expected "[object CSSStyleValue]" but got "[object Undefined]"
Harness: the test ran to completion.
